+++ src/main/kotlin/dev/xeonkryptos/xeonrobotframeworkplugin/ide/execution/ui/editor/RobotTestExecutionMode.kt
@@ -1,0 +1,5 @@
+package dev.xeonkryptos.xeonrobotframeworkplugin.ide.execution.ui.editor
+
+enum class RobotTestExecutionMode(val unitExecutionMode: Boolean) {
+    TEST_CASES(true), TASKS(true), DIRECTORIES(false)
+}